feat!: Add shared components to the napari-copick UI (#11)
Adds gallery and info widget from `copick-shared-ui` 0.1.1 to the
napari-copick interface. Makes napari-copick rely on the same thumbnail
cache as chimerax-copick.

### Gallery widget
The gallery widget provides a searchable list of runs in the copick
config with thumbnails. Thumbnails are generated from the tomogram with
the highest binning, denoised tomograms are preferred. Thumbnails are
stored in a cache directory (different location based on OS), and
retrieved from there if existing.

[Screencast from 2025-07-06
11-07-15.webm](https://github.com/user-attachments/assets/cead52a3-66f5-4a22-8cfe-2ea68fd4d9ca)


### Info widget
The info widget provides and overview of the full contents of a run,
including thumbnails of different tomogram versions.

[Screencast from 2025-07-06
11-15-19.webm](https://github.com/user-attachments/assets/d1c3e7de-e934-44c7-97d9-862fe9bbac57)

### New Object dialog
This dialog allows modifying the currently open copick config to add or
edit object types.

[Screencast from 2025-07-06
11-19-44.webm](https://github.com/user-attachments/assets/847be3a2-b2f1-4806-a855-c0d98ca347cd)